requirements.txt
mysql-connector-python==8.1.0
pymongo==4.5.0
python-dotenv==1.0.1
bcrypt==4.0.1